+++ lib/appconfig.php
@@ -175,6 +175,13 @@ class AppConfig {
 	private $_customizationForcesave = "customizationForcesave";
 
 	/**
+	 * The config key for the live view on share setting
+	 *
+	 * @var string
+	 */
+	private $_liveViewOnShare = "liveViewOnShare";
+
+	/**
 	 * The config key for the help display setting
 	 *
 	 * @var string
@@ -904,6 +911,28 @@ class AppConfig {
 	}
 
 	/**
+	 * Save live view on share setting
+	 *
+	 * @param bool $value - live view on share
+	 * 
+	 * @return void
+	 */
+	public function setLiveViewOnShare($value) {
+		$this->logger->info("Set live view on share: " . json_encode($value), ["app" => $this->appName]);
+
+		$this->config->setAppValue($this->appName, $this->_liveViewOnShare, json_encode($value));
+	}
+
+	/**
+	 * Get live view on share setting
+	 *
+	 * @return bool
+	 */
+	public function getLiveViewOnShare() {
+		return $this->config->getAppValue($this->appName, $this->_liveViewOnShare, "false") === "true";
+	}
+
+	/**
 	 * Save help display setting
 	 *
 	 * @param bool $value - display help
@@ -1233,12 +1262,22 @@ class AppConfig {
 	}
 
 	/**
+	 * Get the error text of the status settings
+	 *
+	 * @return bool
+	 */
+	public function getSettingsError() {
+		return $this->config->getAppValue($this->appName, $this->_settingsError, "");
+	}
+
+
+	/**
 	 * Get the status settings
 	 *
 	 * @return bool
 	 */
 	public function settingsAreSuccessful() {
-		return empty($this->config->getAppValue($this->appName, $this->_settingsError, ""));
+		return empty($this->getSettingsError());
 	}
 
 	/**