+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - there are only system, light and dark themes
 - toolbarNoTabs settings removed
 - fix opening version from sidebar
+- display error when opening settings page if it exists
 
 ## 9.9.1
 ## Added