+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - insert svg to editor
 - view vsdx, vssx, vstx, vsdm, vssm, vstm
 - view odg, md
+- setting for enabling live-view for shared docs
 
 ## Changed
 - edit xlsb