+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-##
+## 9.10.1
 ## Added
 - insert svg to editor
 - view vsdx, vssx, vstx, vsdm, vssm, vstm